Finished ADC base implementation, no testing yet Working on TM4C... .c
TODO: TM4C... .c, i2c.c, timer.c
