+++ hw/meson.build
@@ -68,3 +68,5 @@ subdir('sparc')
 subdir('sparc64')
 subdir('tricore')
 subdir('xtensa')
+
+subdir('tm4c123gh6pm')
