+++ hw/tm4c123gh6pm/Kconfig
@@ -1,0 +1,17 @@
+config TM4C123GH6PM
+    bool
+    default y
+    depends on TCG && ARM
+    imply I2C_DEVICES
+    select ARM_V7M
+    select CMSDK_APB_WATCHDOG
+    select I2C
+    select PL011 # UART
+    select PL022 # SPI
+    select PL061 # GPIO
+    select SSD0303 # OLED display
+    select SSD0323 # OLED display
+    select SSI_SD
+    select STELLARIS_GAMEPAD
+    select STELLARIS_GPTM # general purpose timer module
+    select UNIMP
