+++ hw/tm4c123gh6pm/adc.h
@@ -1,0 +1,6 @@
+#ifndef TM4_ADC_H_
+#define TM4_ADC_H_
+
+static void adc_register_types(void);
+
+#endif
