+++ hw/tm4c123gh6pm/i2c.h
@@ -1,0 +1,6 @@
+#ifndef TM4_I2C_H_
+#define TM4_I2C_H_
+
+static void i2c_register_types(void);
+
+#endif
